+++ packages/vector_graphics_codec/CHANGELOG.md
@@ -1,6 +1,10 @@
 # CHANGELOG
 
-## 1.1.6
+## 1.1.5+1
+
+- Remove/update some invalid assertions related to image formats.
+
+## 1.1.5
 
 - Add support for encoding control points as IEEE 754-2008 half precision
   floating point values.